[Test] Refactored a bit to allow stubbing more than one endpoint.
