+++ WordPress/src/androidTest/java/org/wordpress/android/e2e/StatsTests.kt
@@ -4,7 +4,7 @@ import androidx.test.espresso.Espresso
 import androidx.test.espresso.matcher.ViewMatchers
 import dagger.hilt.android.testing.HiltAndroidTest
 import org.junit.After
-import org.junit.Assume.assumeTrue
+import org.junit.Assume
 import org.junit.Before
 import org.junit.Test
 import org.wordpress.android.BuildConfig
@@ -13,15 +13,14 @@ import org.wordpress.android.e2e.pages.MySitesPage
 import org.wordpress.android.support.BaseTest
 import org.wordpress.android.support.ComposeEspressoLink
 import org.wordpress.android.support.WPSupportUtils
-import org.wordpress.android.util.StatsKeyValueData
-import org.wordpress.android.util.StatsMocksReader
-import org.wordpress.android.util.StatsVisitsData
+import org.wordpress.android.wiremock.WireMockStub
+import org.wordpress.android.wiremock.WireMockUrlPath
 
 @HiltAndroidTest
-class StatsTests : BaseTest() {
+class StatsTests : BaseTest(listOf(WireMockStub(urlPath = WireMockUrlPath.FEATURE_RESPONSE, fileName = "new-stats-feature-response.json"))) {
     @Before
     fun setUp() {
-        assumeTrue(BuildConfig.IS_JETPACK_APP)
+        Assume.assumeTrue(BuildConfig.IS_JETPACK_APP)
         ComposeEspressoLink().unregister()
         logoutIfNecessary()
         wpLogin()
@@ -38,25 +37,9 @@ class StatsTests : BaseTest() {
 
     @Test
     fun e2eAllDayStatsLoad() {
-        val todayVisits = StatsVisitsData("97", "28", "14", "11")
-        val postsList: List<StatsKeyValueData> = StatsMocksReader().readDayTopPostsToList()
-        val referrersList: List<StatsKeyValueData> = StatsMocksReader().readDayTopReferrersToList()
-        val clicksList: List<StatsKeyValueData> = StatsMocksReader().readDayClicksToList()
-        val authorsList: List<StatsKeyValueData> = StatsMocksReader().readDayAuthorsToList()
-        val countriesList: List<StatsKeyValueData> = StatsMocksReader().readDayCountriesToList()
-        val videosList: List<StatsKeyValueData> = StatsMocksReader().readDayVideoPlaysToList()
-        val downloadsList: List<StatsKeyValueData> = StatsMocksReader().readDayFileDownloadsToList()
         MySitesPage()
             .go()
             .goToStats()
-            .openDayStats()
-            .assertVisits(todayVisits)
-            .scrollToPosts().assertPosts(postsList)
-            .scrollToReferrers().assertReferrers(referrersList)
-            .scrollToClicks().assertClicks(clicksList)
-            .scrollToAuthors().assertAuthors(authorsList)
-            .scrollToCountries().assertCountries(countriesList)
-            .scrollToVideos().assertVideos(videosList)
-            .scrollToFileDownloads().assertDownloads(downloadsList)
+            .hasNewStatTabs()
     }
 }
